440 SessionTimeout: Client Session Expired Must Log In Again
Usage is not widespread as this is a code used by Microsoft.
https://stackoverflow.com/questions/1653493/what-http-status-code-is-supposed-to-be-used-to-tell-the-client-the-session-has/22484262#22484262

However given the frequent usage of Promises with tink,
it might seem a good idea to have this error code for better discrimination.
For instance if one was to sanitize http error messages in Production to
apply those from the error code, having a mere 401 would be problematic.

With a 440 there is not this problem anymore.